tests/chrome_dinosaur_game/requirements.txt
black==22.6.0
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
docutils==0.19
freetype-py==2.3.0
glfw==2.5.4
glm==0.4.3
idna==3.3
importlib-metadata==4.12.0
keyboard==0.13.5
keyring==23.8.2
mypy-extensions==0.4.3
Nuitka==1.0.3
numpy==1.23.1
ordered-set==4.1.0
pathspec==0.9.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
Pxucz==0.0.1.8
PyAudio==0.2.12
Pygments==2.12.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
pywin32-ctypes==0.2.0
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
shiboken6==6.3.1
six==1.16.0
tomli==2.0.1
twine==4.0.1
urllib3==1.26.11
webencodings==0.5.1
zipp==3.8.1
